Proper fix for textbox UART hang/crash
THANK YOU @0xchocolate

+++ applications/services/gui/modules/text_box.c
@@ -104,9 +104,7 @@ static void text_box_insert_endline(Canvas* canvas, TextBoxModel* model) {
     if(model->focus == TextBoxFocusEnd && line_num > lines_on_screen) {
         // Set text position to 5th line from the end
         const char* end = model->text + furi_string_size(model->text_formatted);
-        // TODO: Find proper fix, this prevents BusFault but hangs GUI after another while? makes no sense
-        for(uint8_t i = 0; i < line_num - lines_on_screen; i++) {
-            // Debugger shows it's stuck in this loop, but it also shows the while condition is false??
+        for(size_t i = 0; i < line_num - lines_on_screen; i++) {
             while(model->text_pos < end) {
                 if(*model->text_pos++ == '\n') break;
             }
